Conflicts:
	README.md

+++ src/main/java/io/sporkpgm/map/SporkMap.java
@@ -3,11 +3,11 @@ package io.sporkpgm.map;
 import io.sporkpgm.ListenerHandler;
 import io.sporkpgm.map.generator.NullChunkGenerator;
 import io.sporkpgm.module.Module;
+import io.sporkpgm.module.ObjectiveModule;
 import io.sporkpgm.module.modules.filter.FilterCollection;
 import io.sporkpgm.match.Match;
 import io.sporkpgm.module.ModuleCollection;
 import io.sporkpgm.module.builder.BuilderContext;
-import io.sporkpgm.module.builder.BuilderFactory;
 import io.sporkpgm.module.modules.info.Contributor;
 import io.sporkpgm.module.modules.kits.KitModule;
 import io.sporkpgm.module.modules.team.TeamCollection;
@@ -16,10 +16,13 @@ import io.sporkpgm.module.modules.region.RegionCollection;
 import io.sporkpgm.scoreboard.DefaultScoreboard;
 import io.sporkpgm.scoreboard.ScoreboardHandler;
 import io.sporkpgm.scoreboard.exceptions.IllegalScoreboardException;
+import io.sporkpgm.scoreboard.objective.ObjectiveScoreboard;
 import io.sporkpgm.user.User;
+import io.sporkpgm.util.ClassUtils;
 import io.sporkpgm.util.FileUtil;
 import io.sporkpgm.util.Log;
 import io.sporkpgm.util.SporkConfig.Settings;
+import io.sporkpgm.win.WinConditionSet;
 import org.bukkit.Bukkit;
 import org.bukkit.World;
 import org.bukkit.WorldCreator;
@@ -27,7 +30,9 @@ import org.bukkit.event.Listener;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public class SporkMap {
 
@@ -40,21 +45,38 @@ public class SporkMap {
 	private ScoreboardHandler scoreboard;
 
 	private World world;
+	private boolean ended;
+	private TeamModule winner;
+	private Map<TeamModule, WinConditionSet> conditions;
 
 	public SporkMap(SporkLoader loader) {
 		this.loader = loader;
+		this.scoreboard = new ScoreboardHandler(this);
 		this.modules = loader.getModules().clone(this);
 		this.modules.add(new BuilderContext(this, loader, loader.getDocument()));
 		this.regions = new RegionCollection(this);
 		this.filters = new FilterCollection(this);
-		this.teams = new TeamCollection(this, modules.getModules(TeamModule.class));
-		this.scoreboard = new ScoreboardHandler(this);
 
 		try {
 			DefaultScoreboard board = this.scoreboard.get(DefaultScoreboard.class);
 			this.scoreboard.setMain(board);
 		} catch(IllegalScoreboardException e) {
 			Log.exception(e);
+		}
+
+		this.scoreboard.register(teams.getObservers());
+		for(TeamModule team : teams.getTeams()) {
+			this.scoreboard.register(team);
+		}
+
+		if(modules.getModules(ObjectiveModule.class).size() > 0) {
+			try {
+				ObjectiveScoreboard board = this.scoreboard.get(ObjectiveScoreboard.class);
+				board.setup();
+				this.scoreboard.setMain(board);
+			} catch(IllegalScoreboardException e) {
+				Log.exception(e);
+			}
 		}
 	}
 
@@ -99,6 +121,10 @@ public class SporkMap {
 	}
 
 	public TeamCollection getTeams() {
+		if(teams == null) {
+			this.teams = new TeamCollection(this, modules.getModules(TeamModule.class));
+		}
+
 		return teams;
 	}
 
@@ -114,13 +140,63 @@ public class SporkMap {
 		return world;
 	}
 
+	public void setEnded(boolean ended) {
+		this.ended = ended;
+
+		if(this.winner == null) {
+			this.winner = getTeams().getObservers();
+		}
+	}
+
+	public boolean hasEnded() {
+		return ended;
+	}
+
+	public void checkEnded() {
+		if(!ended && getWinner() != null && !getWinner().isObservers()) {
+			setEnded(true);
+		}
+	}
+
 	public TeamModule getWinner() {
-		return null;
+		for(TeamModule module : getConditions().keySet()) {
+			WinConditionSet set = conditions.get(module);
+			if(set.isCompleted()) {
+				return module;
+			}
+		}
+
+		return getTeams().getObservers();
+	}
+
+	public void setWinner(TeamModule winner) {
+		this.winner = winner;
+	}
+
+	public Map<TeamModule, WinConditionSet> getConditions() {
+		if(conditions == null) {
+			this.conditions = new HashMap<>();
+			for(TeamModule team : teams.getTeams()) {
+				this.conditions.put(team, new WinConditionSet());
+			}
+		}
+
+		return conditions;
+	}
+
+	public WinConditionSet getConditionSet(TeamModule team) {
+		return getConditions().get(team);
 	}
 
 	public void start() {
 		for(Module module : modules.getModules()) {
 			module.start();
+		}
+	}
+
+	public void stop() {
+		for(Module module : modules.getModules()) {
+			module.stop();
 		}
 	}
 
@@ -139,6 +215,7 @@ public class SporkMap {
 			return false;
 		}
 
+		Log.info("World for " + toString() + " is " + world.getName());
 		return true;
 	}
 
@@ -151,7 +228,10 @@ public class SporkMap {
 
 		this.world = null;
 		String name = Settings.prefix() + match.getID();
-		Bukkit.unloadWorld(name, false);
+		boolean success = Bukkit.unloadWorld(name, false);
+		if(!success) {
+			return false;
+		}
 
 		File dest = new File(Bukkit.getWorldContainer(), name);
 		FileUtil.delete(dest);
